Move locations layers to different components
